refactor: replace AnimationOptions with $Transition type in motion package
